+++ crates/fuel-gas-price-algorithm/gas-price-analysis/.gitignore
@@ -1,2 +1,3 @@
 Cargo.lock
 gas-prices.png
+/predefined_data
