+++ crates/fuel-gas-price-algorithm/gas-price-analysis/src/charts.rs
@@ -1,5 +1,52 @@
 use super::*;
 
+pub fn draw_chart(results: SimulationResults, p_comp: i64, d_comp: i64, file_path: &str) {
+    let SimulationResults {
+        gas_prices,
+        exec_gas_prices,
+        da_gas_prices,
+        fullness,
+        bytes_and_costs,
+        actual_profit,
+        projected_profit,
+        pessimistic_costs,
+    } = results;
+
+    let max_actual_profit = pretty(*actual_profit.iter().max().unwrap() as u64);
+    println!("max_actual: {max_actual_profit}");
+
+    let plot_width = 640 * 2 * 2;
+    let plot_height = 480 * 3;
+
+    let root =
+        BitMapBackend::new(file_path, (plot_width, plot_height)).into_drawing_area();
+    root.fill(&WHITE).unwrap();
+    let (window_one, lower) = root.split_vertically(plot_height / 4);
+    let (window_two, new_lower) = lower.split_vertically(plot_height / 4);
+    let (window_three, window_four) = new_lower.split_vertically(plot_height / 4);
+
+    draw_fullness(&window_one, &fullness, "Fullness");
+
+    draw_bytes_and_cost_per_block(&window_two, &bytes_and_costs, "Bytes Per Block");
+
+    draw_profit(
+        &window_three,
+        &actual_profit,
+        &projected_profit,
+        &pessimistic_costs,
+        &format!("Profit p_comp: {p_comp:?}, d_comp: {d_comp:?}"),
+    );
+    draw_gas_prices(
+        &window_four,
+        &gas_prices,
+        &exec_gas_prices,
+        &da_gas_prices,
+        "Gas Prices",
+    );
+
+    root.present().unwrap();
+}
+
 pub fn draw_gas_prices(
     drawing_area: &DrawingArea<BitMapBackend, Shift>,
     gas_prices: &[u64],
