+++ crates/fuel-gas-price-algorithm/gas-price-analysis/src/main.rs
@@ -1,4 +1,6 @@
-use crate::simulation::GeneratedDaCost;
+use crate::charts::draw_chart;
+use crate::simulation::get_da_cost_per_byte_from_source;
+use crate::simulation::{arbitrary_cost_per_byte,};
 use plotters::prelude::*;
 use rand::{
     rngs::StdRng,
@@ -27,17 +29,6 @@ mod simulation;
 
 mod charts;
 
-pub fn pretty<T: ToString>(input: T) -> String {
-    input
-        .to_string()
-        .as_bytes()
-        .rchunks(3)
-        .rev()
-        .map(std::str::from_utf8)
-        .collect::<Result<Vec<&str>, _>>()
-        .unwrap()
-        .join(",") // separator
-}
 
 
 use clap::{Parser, Subcommand};
@@ -76,7 +67,8 @@ enum Mode {
 
 #[derive(Subcommand)]
 enum Source {
-    Generated { size: usize }
+    Generated { size: usize },
+    Predefined { file_path: String }
 }
 
 fn main() {
@@ -86,18 +78,18 @@ fn main() {
 
     let (results, (p_comp, d_comp)) = match args.mode {
         Mode::WithValues { p, d, source } => {
-            let Source::Generated { size } = source;
-            let da_cost_source = GeneratedDaCost { size, update_period: UPDATE_PERIOD };
+            let da_cost_per_byte = get_da_cost_per_byte_from_source(source, UPDATE_PERIOD);
+            let size = da_cost_per_byte.len();
             println!("Running simulation with P: {}, D: {}, and {} blocks", pretty(p), pretty(d), pretty(size));
-            let simulator = Simulator::new(da_cost_source);
+            let simulator = Simulator::new(da_cost_per_byte);
             let result = simulator.run_simulation(p, d, UPDATE_PERIOD);
             (result, (p, d))
         },
         Mode::Optimization { iterations, source} => {
-            let Source::Generated { size } = source;
+            let da_cost_per_byte = get_da_cost_per_byte_from_source(source, UPDATE_PERIOD);
+            let size = da_cost_per_byte.len();
             println!("Running optimization with {iterations} iterations and {size} blocks");
-            let da_cost_source = GeneratedDaCost { size, update_period: UPDATE_PERIOD };
-            let simulator = Simulator::new(da_cost_source);
+            let simulator = Simulator::new(da_cost_per_byte);
             let (results, (p, d)) = naive_optimisation(simulator, iterations as usize, UPDATE_PERIOD);
             println!("Optimization results: P: {}, D: {}", pretty(p), pretty(d));
             (results, (p, d))
@@ -110,50 +102,14 @@ fn main() {
     }
 }
 
-fn draw_chart(results: SimulationResults, p_comp: i64, d_comp: i64, file_path: &str) {
-    let SimulationResults {
-        gas_prices,
-        exec_gas_prices,
-        da_gas_prices,
-        fullness,
-        bytes_and_costs,
-        actual_profit,
-        projected_profit,
-        pessimistic_costs,
-    } = results;
-
-    let max_actual_profit = pretty(*actual_profit.iter().max().unwrap() as u64);
-    println!("max_actual: {max_actual_profit}");
-
-    let plot_width = 640 * 2 * 2;
-    let plot_height = 480 * 3;
-
-    let root =
-        BitMapBackend::new(file_path, (plot_width, plot_height)).into_drawing_area();
-    root.fill(&WHITE).unwrap();
-    let (window_one, lower) = root.split_vertically(plot_height / 4);
-    let (window_two, new_lower) = lower.split_vertically(plot_height / 4);
-    let (window_three, window_four) = new_lower.split_vertically(plot_height / 4);
-
-    draw_fullness(&window_one, &fullness, "Fullness");
-
-    draw_bytes_and_cost_per_block(&window_two, &bytes_and_costs, "Bytes Per Block");
-
-    draw_profit(
-        &window_three,
-        &actual_profit,
-        &projected_profit,
-        &pessimistic_costs,
-        &format!("Profit p_comp: {p_comp:?}, d_comp: {d_comp:?}"),
-    );
-    draw_gas_prices(
-        &window_four,
-        &gas_prices,
-        &exec_gas_prices,
-        &da_gas_prices,
-        "Gas Prices",
-    );
-
-    root.present().unwrap();
+pub fn pretty<T: ToString>(input: T) -> String {
+    input
+        .to_string()
+        .as_bytes()
+        .rchunks(3)
+        .rev()
+        .map(std::str::from_utf8)
+        .collect::<Result<Vec<&str>, _>>()
+        .unwrap()
+        .join(",") // separator
 }
-
